refactor: Remove old github
