--- py-scripts/toolbelt/github/__init__.py
@@ -1,3 +1,0 @@
-from .commit import commit_manifests
-
-__all__ = ["commit_manifests"]
